% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FL.R
\docType{data}
\name{FL}
\alias{FL}
\title{Data of INSURANCE}
\format{
A tibble with 36,633 rows and 15 variables:
\describe{
\item{policyID}{dbl IDS of the policies}
\item{statecode}{chr state code of the insurance data}
\item{county}{chr county of the insurance}
\item{\code{EA}}{dbl denoting EA class}
\item{\code{HU}}{dbl denoting HU class}
\item{\code{FL}}{dbl denoting FL class}
\item{\code{FR}}{dbl denoting FR class}
\item{\code{2011}}{dbl denoting year class}
\item{\code{2012}}{dbl denoting year class}
\item{\code{EQ}}{dbl denoting EQ class}
\item{\code{HU}}{dbl denoting HU class}
\item{\code{FL}}{dbl denoting FL class}
\item{\code{FR}}{dbl denoting FR class}
\item{\code{LATI}}{dbl denoting Latitude class}
\item{\code{LONG}}{dbl denoting longitude class}
}
}
\source{
\url{https://support.spatialkey.com/spatialkey-sample-csv-data/}
}
\usage{
FL
}
\description{
The sample insurance file contains 36,634 records in Florida for 2012 from a sample company that implemented an agressive growth plan in 2012.  There are total insured value (TIV) columns containing TIV from 2011 and 2012
}
\keyword{datasets}
